requirements.txt
groq
PyYAML
streamlit